FIX - increment slab allocator total (#363)
Credit @andrew-lee-architect for 30m huddle / discussion on allocator
interworkings.

Here is a one liner that should fix this issue:
[A-237](https://linear.app/architect-xyz/issue/A-237/kraken-website-order-on-broke-something-in-canary)

# Preface:
The StaticBumpAllocator, defined in `sdk/src/symbology/allocator.rs`.
Allocates "slabs", which are arrays of a fixed size. It manually tracks
the position to next position to insert at using the pos field, which is
a pointer.

### Allocator Insert
The allocator is singly linked list that points backwards to last
allocated slab. When the slab is full, the pos pointer is reset and a
new slab is allocated (this happens in insert) and we assign the
allocator prev pointer to the last static allocator

### Allocator Snapshot
When you create a snapshot using the allocator it iterates backwards
through each slab, creating an array of pointers to each slab data. Then
it loops through each slab calling a function you pass.

# Problem
The total was never incrementing on inserts to allocator. Looking at the
static_ref file which helps applications use the allocator:

In `sdk/src/symbology/static_ref.rs`:
```rust
pub trait StaticRef<T: Symbolic, const SLAB_SIZE: usize>:
    Clone + Copy + Deref<Target = T> + 'static
{
    /// Returns a reference to the global pool of T's
    fn allocator() -> &'static Mutex<StaticBumpAllocator<T, SLAB_SIZE>>;

    fn allocator_counter() -> &'static AtomicUsize;

    /// Returns a snapshot of the current global pool of T's
    fn allocator_snapshot<F: FnMut(Self)>(
        prev: Option<AllocatorSnapshot<T>>,
        mut new: F,
    ) -> AllocatorSnapshot<T> {
        match prev {
            None => {
                Self::allocator().lock().snapshot(prev, |i| new(Self::from_pointee(i)))
            }
            Some(p) => {
                if p.total < Self::allocator_counter().load(Ordering::Relaxed) {
                    Self::allocator()
                        .lock()
                        .snapshot(Some(p), |i| new(Self::from_pointee(i)))
                } else {
                    p
                }
            }
        }
    }

...
```
prev is the pointer in the linked list to last element. So the first
time this will always succeed and call snapshot. However, the next time
we update it will just return p as total is stuck at 0. Note that
Self::allocator_counter() is just a ref to the total of the allocator
(which again is stuck at 0)

**The solution is the one line change which increments total so as to
keep track of all allocated elements in the slab allocator**

This fix will allow counterparties to subsequently update past the first
update (which caused the linked issue)

Note as Andrew pointed out, there may be some thread safety concerns on
insert when allocator is not wrapped by mutex