Channel topic subscriptions (#411)
Introduces the notion of message topics, which are a property of
`TypedMessage`s. External clients who subscribe using the Rust SDK can
request to subscribe to topics, which will allow them to see messages
not directly addressed to them.

The first use case will be external orderflow from Binance, where
external orders and fills have no reply_to chain that leads back to any
particular client, because the order wasn't initiated from Architect to
begin with. This particular implementation is done in a separate branch
for review clarity.