Delayed republisher for marketdata (#556)
# Description
This is a republishing feed that consumes some upstream feed and then
publishes it at some delay. This is motivated by us wanting to publish a
CME feed at a 15m delay.

## Design
The way this feed works is that it maintains its own set of BookClients
and then publishes a NetidxFeed based off of that. It's very similar to
(and a bit copied from) ManagedMarketdata in that it subscribes to the
update events from upstream feeds and then updates the BookClients based
off of that. The main difference is that it queues the update events up
and only shows them to the BookClients after the configured delay.

The original way I approached this was by using ManagedMarketdata and
then periodically taking snapshots of the book and periodically
republishing it. The downside was that the data wouldn't be very
granular, which I'm still not sure is that bad if you're already looking
at 15m delayed feed.

Anticipating some conversation around:
- What should the publishing path be? Right now the way I've chosen to
do it is a bit hacky, I have the netdix feed routes set to e.g.
"DATABENTO-DELAY".
- This currently doesn't republish the last information. Does it need
to?
- I'm not sure I fully understand the snapshot issue with Brett's
republisher, but I believe the way that these BookClients work is that
they write an empty value to request a snapshot when they start up, but
we have control over when the BookClient processes the snapshot since
snapshots are included on the event stream along with the individual
updates.

## Testing
Spun up a delayed DATABENTO feed for local testing.

If you want to reproduce the testing environment:
- Add CME/DATABENTO-DELAY to use_local_marketdata in your architect
config.
- Here's a delayed config:
```
venue: "CME"
route: "DATABENTO"
delay: "2m"
snapshot_interval: "30s" 
```
- Run this`$ ./target/debug/architect marketdata feed delayed
var/delayed_cme.yml`

## Changelog

- Creates a new delayed republisher for marketdata

## Semver

Please check the box below to let the release bot know how to increment
the version number.

- [ ] This PR includes breaking changes to `api`, `sdk`, `TypedMessage`,
component configs, component state, etc.
- [ ] This PR includes backwards-compatible changes to `api`, `sdk`,
`TypedMessage`, component configs, component state,
  etc.

---------

Co-authored-by: Andrew Lee <126016409+andrew-lee-architect@users.noreply.github.com>