L2 marketdata publishing via grpc  (#724)
# Description
This PR implements L2 marketdata publishing via grpc and as a
pre-requisite, does some rearranging of code inside of
`MultiConnectionFeed`.

### L2 marketdata endpoints
The L2 publishing is actually split into 2 separate endpoints: one to
get a snapshot and one to subscribe to diffs. The client is responsible
for using the 2 endpoints to construct a coherent state for the book.
I've written a small client library in debug to do exactly this.

### Feed shards
There's a bit of moving various pieces of code around but main thing
that has changed is that the individual "tasks" as they used to be
called are now called "shards" and has been factored into a struct in a
separate file. These shards can be created, run, and owns the feed state
and the main channels of L1 and L2 updates. It then exposes a view of
the internal state and update channels. When reading about design
patterns in rust, I found that many call this the "Actor" pattern (see
[this](https://ryhl.io/blog/actors-with-tokio/) and
[this](https://medium.com/@maturationofthe/leveraging-rusts-tokio-library-for-asynchronous-actor-model-cf6d477afb19)).

I ended up choosing to use a concurrent data structure that the rpc
service can have access to instead of relying exclusively on message
passing between shard <--> rpc service because I wanted the shards to be
focused on subscribing to marketdata and maintaining book state, not
with serving every rpc with a fresh snapshot.

## Testing 
Tested by spinning up a local OKX feed and subscribing to the L2 stream.
The L2 client subscription looked like it was working as expected.

I also ran the feed differ, which checked the bbo at various increments.
I tried it every 1s and every 5s. For 5s increments, the grpc feed was
always behind the netidx feed by about 30ms. For 1s increments, the grpc
feed was usually behind the netidx feed by about 30ms but every once in
a while there would be a race where it would be a bit further ahead. In
both, the bbo rarely every differed after running for a few minutes.

## Changelog

- [x] Update `CHANGELOG.md` to reflect the changes in this PR

---------

Co-authored-by: Jen Cecelia Reiss <jen@architect.xyz>
Co-authored-by: Andrew Lee <126016409+andrew-lee-architect@users.noreply.github.com>