ManagedMarketdata continues processing events after an error (#775)
# Description
This was the easiest change to make in order to let ManagedMarketdata
keep processing events for other books after one errors, but it's a bit
suboptimal because it could cause resubscriptions with errors in a tight
loop and the book client of the book that had an error will just remain
stale since there's not really a way to mark it as in error. I think we
should consider this case in the L2Client, I'll take a look at it later.

## Alternative considered: cooldown for resubscription after error
I did try an alternative implementation first to keep track of
MarketRefs that have books in error and have some sort of cooldown
period before allowing resubscription, but that would mean that
[subscribe] would have to return a [Result] which not all callsites
would be able to handle gracefully without more refactoring. So this is
a reasonable middle ground in my opinion.

## Changelog

- [x] Update `CHANGELOG.md` to reflect the changes in this PR